+++ semcore/data-table/src/components/DataTable/DataTable.tsx
@@ -92,6 +92,7 @@ class DataTableRoot<
   private tableRef = React.createRef<HTMLDivElement>();
   private headerRef = React.createRef<HTMLDivElement>();
   private spinnerRef = React.createRef<HTMLDivElement>();
+  private containerResizeEndTimeoutId: ReturnType<typeof setTimeout> | null = null;
 
   private gridAreaGroupMap = new Map<number, string>();
 
@@ -759,6 +760,14 @@ class DataTableRoot<
     this.changeFocusCell(-1, cellIndex === -1 ? 0 : cellIndex, 'up');
   };
 
+  handleContainerResizeEnd = () => {
+    if (this.containerResizeEndTimeoutId) {
+      clearTimeout(this.containerResizeEndTimeoutId);
+    }
+
+    this.containerResizeEndTimeoutId = setTimeout(this.calculateVerticalShadow, 0);
+  };
+
   render() {
     const SDataTable = Root;
     const {
@@ -815,6 +824,7 @@ class DataTableRoot<
         styles={scrollStyles}
         onScroll={this.handleScroll}
         disableAutofocusToContent={true}
+        onResize={this.handleContainerResizeEnd}
       >
         <ScrollArea.Container
           tabIndex={-1}
@@ -1219,8 +1229,6 @@ class DataTableRoot<
         groupIndex++;
       }
     });
-
-    console.log(calculatedColumns, treeColumns);
 
     return [calculatedColumns, treeColumns];
   }